Converted commandHandler into ClientCommandHandler class.
(cherry picked from commit 2f95bcb2887b0928fe0c284491ef75724d0a344b)
