+++ Classes/ClientEvents/MessageCreate.ts
@@ -4,7 +4,6 @@
 
 import { ChannelType, Events } from "discord.js";
 import ClientEvent from "../ClientEvent.ts";
-import { executeCommand } from "../../Modules/commandHandler.ts";
 import { processIncomingMessage } from "../../Modules/messageHandler.ts";
 
 export default new ClientEvent({
@@ -25,7 +24,7 @@ export default new ClientEvent({
         let isCommand = messageStartsWithCommandAlias || message.channel.type === ChannelType.DM || message.channel.id === game.guildContext.commandChannel.id;
         if (isCommand) {
             const command = messageStartsWithCommandAlias ? message.content.substring(game.settings.commandPrefix.length) : message.content;
-            isCommand = await executeCommand(command, game, message);
+            isCommand = await game.clientContext.commandHandler.executeCommand(command, game, message);
         }
         if (message.channel.type !== ChannelType.DM && !isCommand && game.inProgress) {
             processIncomingMessage(game, message);
