+++ Docs/reference/commands/index.md
@@ -8,6 +8,6 @@ Commands are loaded from the commands directory when Alter Ego is booted up. Eac
 message which sent the command and carry out the desired behavior.
 
 All commands are passed through
-the [commandHandler module](https://github.com/MsVBLANK/Alter-Ego/blob/master/Modules/commandHandler.ts) before being
-executed. The purpose ouf this module is to determine who is sending the command, and if they have permission to do so.
-All commands are restricted to a single permission level based on the sender's Discord roles in the game server.
+the [ClientCommandHandler](https://github.com/MsVBLANK/Alter-Ego/blob/master/Classes/ClientCommandHandler.ts) before
+being executed. The purpose ouf this module is to determine who is sending the command, and if they have permission to
+do so. All commands are restricted to a single permission level based on the sender's Discord roles in the game server.
